Rename virtual_fs.rs to file_system.rs
